Impl proone-htbtclient, bugfix, --enable-mttools
* Remove --enable-hostinfod and add --enable-mttools, which enables
  hostinfod and htbtclient
* Change PRNE_HTBT_PROTO_PORT so that the macro can be used in another
  macro expression(for help message)
* Add prne_mbedtls_perror()
* proone-htbtclient: implement hostinfo command
* proone-htbthost
  * Add --hostcred option
  * Response with status frame on ENOMEM
* htbt: add prne_htbt_serrc_tostr() and prne_htbt_gen_msgid()
* Add prne_ctoupper(), prne_ctolower(), prne_cisspace(), prne_cisprint()
  * The added functions are the locale-independent equivalent of their
    POSIX counterparts
  * Remove Use of the POSIX functions
  * Add test unit

+++ src/htbt.c
@@ -26,7 +26,7 @@
 // between 30 minutes and an hour
 #define HTBT_CNCP_INT_MIN	1800000 // half an hour minimum interval
 #define HTBT_CNCP_INT_JIT	1800000 // half an hour jitter
-#define HTBT_LBD_PORT			prne_htobe16(PRNE_HTBT_PROTO_PORT)
+#define HTBT_LBD_PORT			prne_htobe16((uint16_t)PRNE_HTBT_PROTO_PORT)
 #define HTBT_LBD_BACKLOG		4
 #define HTBT_LBD_MAX_CLIENTS	5
